// src/enum-types.ts
export enum EJobFinderRole {
  EMPLOYEES = 'jb-employees',
  EMPLOYERS = 'jb-employers',
}

export enum ECvTemplate {
  TEMPLATE1 = 'cvTemplate1',
  TEMPLATE2 = 'cvTemplate2',
}

export enum EIdentifyType {
  CREDENTIAL = 'credential',
  GOOGLE = 'google',
  FACEBOOK = 'facebook',
}

export enum EApplicationStatus {
  PENDING = 'pending',
  ACCEPTED = 'accepted',
  REJECTED = 'rejected',
}